requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
celery
redis
playwright
requests
